Swap width and height in TouchVideoFrame
Currently in TouchVideoFrame and related code, the field 'width'
precedes the field 'height'.
But these fields should be  interpreted as follows:
width -> number  of columns
height -> number of rows
In most notations in mathematics, the matrix sizes are designated as
'mxn', where m is the number of rows and n is the number of columns.

So we make TouchVideoFrame consistent with this, and swap these 2
fields.

Test: atest libinput_tests
The actual test is added in the next commit.
Bug: 123241238

Change-Id: I808e7f354bd7b62d5599324eef205bf4450a91c1