Use InputDeviceIdentifier.getCanonicalName
Currently, there are 2 places with identical logic in InputDevice.cpp.
This is because the file name adjustment to only contain
alphanumerics, dash, and underscore is done for virtual key maps and for
other devices together.
Use the recently added api to simplify this. Also, this will prevent
having to iterate through Vendor_*_Product_* filenames, since those are
already guaranteed to be compliant.

Bug: 113575658
Test: atest libinput_tests inputflinger_tests
Change-Id: Icb22f6e76b83b6d7d170b9e537787f0455db1c1f