SurfaceFlinger: fix sysprop API condition for color management
Device color management can be set via a dedicated property, or inferred
from two other properties. Currently all 3 properties need to be
specified to *some* value in order for color management to be set as
intended. This fix changes this logic so that if one of the properties
is not specified, the default value for color management is used.

Fixes: 126907354
Test: manual by omitting some of the relevant props from device
makefiles; and atest libsurfaceflinger_unittest

Change-Id: Ic3f76a3e50914906cfd53ec8242dedc2d1585186