Rename functions in InputDispatcher
Some of the functions in InputDispatcher are labeled as '*Locked', when
in fact they do not need the caller to hold the lock.
Rename the functions to match the actual needs.

Also, one of the functions does nothing (just return false), but is
labeled as *LockedInterruptible. In this case, it is just done for the
consistency. Keep it and add a REQUIRES annotation.

Test: interact with booted device
Bug: 123097103
Change-Id: I946148f07def7cde35eec93290dc3d54844eef67