SF: Move OutputLayer creation to computeVisibleRegions
This moves part of rebuildLayerStacks to computeVisibleRegions. In
particular it eliminates a second loop over all front-end layers to
identify the ones that are visible on an output. The first loop
computes the visibility, and can immediately create the output layers.

This is a bit of nontrivial restructuring prior to moving the code to
CompositionEngine, since that will also break things up a bit.

Test: atest libsurfaceflinger_unittest libcompositionengine_test
Bug: 121291683
Change-Id: I8a498195f9b5f976e7a69de717818b57494e2f2b