[RoundedCorner] Use the correct source bounds and crop.
Previously we call getBounds to calculate the rounded corner, but that's wrong.
For an app layer, what we really need is the source bounds of the layer, and
the actual crop with transformation but without any parent bounds and display
bounds. However, only the root view has the correct crop information. And thus
this patch passes the correct bounds up to the root, and use the crop from the
root to calculate the rounded corner bounds.

BUG: 125916918
Test: Build, flash and boot. Verify with window transition.
Change-Id: I6c1b92a2c684da956b284b190e8b8a1b45494140