Mark libandroid_runtime_lazy as double_loadable
Because of following two conditions,
- libmediandk(LL-NDK) depends on it.
- libandroid_runtime_lazy is vendor_available:true.

This lib is vendor_available but actually it is not supposed
to be used by vendor processes (which are not app_process and
there is no android_runtime).

If the vendor variant of this lib is loaded
and the delegate functions are called, then dlopen will fail as
expected.

Bug: 125550121
Test: m -j
Change-Id: I7f417d771ccff0679018f9533a5bf1c2b3b43f00
Merged-In: I7f417d771ccff0679018f9533a5bf1c2b3b43f00