Revert "[Rounded Corner] Handle ContainerLayer correctly."
This reverts commit 05eb5c476f386e1629f1a8dd1514a7c945eb30c1.

Reason for revert: Results in weird crop issues
Below is original commit message:
Previously we always use the source bounds of the child layer, this is correct
until two layers overlapped and form the hierachical structure. In this case we
need to use the source bounds of the parent, however, source bounds on
ContainerLayer usually doesn't make sense, in that case we need to use the
child source bounds.

BUG: b/125916918, b/129062310
Test: Test: Build, flash and boot. Verify with window transition for Youtube and Chrome.

BUG: b/129540451
Change-Id: I0d0bfa45082e5513840555da5575b1a49521905f