Avoid sending powerhint right after GPU composition.
Previously we turned off expensive rendering right after we finish GPU
composition, howerver, this will result in SurfaceFlinger keeps sending
expensive rendering hint to power hal and it has a cost. We would like to avoid
that by only sending expensive rendering hint on to power hal when we have GPU
composition and sending it off when we are sure next frame doesn't have GPU
composition.

BUG: 130030334
Test: Verify with WCG Photos
Change-Id: Ib2eaf5a866aa236c9c3bc26bebdd5e60e2391ea6