ASurfaceControl: support framebuffer devices
There are Android devices that use a hwc2onfbadapter that allows
them to ship framebuffer devices but still pass VTS and appear to
the system as if they have hwc2.x*.

Unfortunately, these devices do not support a present fence.
This patch updates the documentation to let developers know that
not all devices will return a present fence.

Test: ASurfaceControlTest
Bug: 129880031
Change-Id: I81dd21dc088b1d2d73d1bdb9c46216ef2e171063