Change the scope to system
The ro.surface_flinger properties are defined in the vendor and can be
used in the vendor. If the scope is internal, the vendor must read the
value using the getproperty function, not the sysprop interface. It is
not the intent of sysprop design. so change the scope to system so that
it can be used by the vendor as well.

Bug: 112386364
Test: m -j && make checkapi
Change-Id: I7a543877eb5c63bab02fa25ef3954456c144ce6b