[SurfaceFlinger] Use SDR data space for HDR video if it's in GPU composition.
If hardware composer supports HDR layer, then this will likely happen in window
transition and rotation, etc, and blending might happen on HDR layer. However,
blending in PQ gamma space is not nice, and thus we pick the widest SDR data
space for HDR video if it's in GPU composition.

As a result, color shift will be observed until RenderEngine can align with the
behaviour of hardware composer.

BUG: 130400004
Test: Verify with Youtube HDR rotation.
Change-Id: I3f517205fa319fe7c8f003f5997a4fe97be9b214