add a static lib of core bufferqueue logic for media
Add a static lib consisting of only the core bufferqueue
logic, which media codec services can use wihtout pulling
the rest of the lib into the apex (and also the runtime
apex namespace). This helps to reduce both the apex size
and the runtime memory of the mediaswcodec.

bug: 128894663
test: build; libgui.so doesn't appear in the apex;
mediaswcodec service loaded correctly at boot

Change-Id: I129790ca7bcae2829b737a96c7465c8f2c5d17eb