AppOpsCallback: do not log error on null package name
Java side can pass 'null' as packageName parameter for
IAppOpsCallback::opChanged, e.g. when calling it via
AppOpsService.notifyWatchersOfChange.

BnAppOpsCallback on the native side should not emit
"Reading a NULL string not supported here" error logs in
this case. For that, a different overload of
'Parcel::readString16' method must be used.

Bug: 130038586
Test: turn DND on / off, watch logcat
Change-Id: I69d28d6a998d75acf987bf49fb0cbfb75cb95828