Use bitset api in InputClassifier
We are currently doing some custom manipulations with bits in
InputClassifier. But the manipulations are making incorrect assumptions
about the behaviour of the bitset apis.

To maximize consistency, extend the use of bitset apis to
InputClassifier. This allows the same apis to be used in the HAL in
order to read the axis values.

Bug: 131927078
Test: atest inputflinger_tests

Change-Id: Icab062c695ca07d5cc404082d3de7b334cbc0e19