Remove 'FMQ wake without any events' log
This log line was causing signficant spam whenever sensors were
active. The original intent of the log line was to identify if the
Sensors HAL incorrectly woke the framework when no sensor events were
available. However, in practice, the framework is able to read events
from the FMQ before the EventFlag is notified which results in the log
line being generated. In its current form, the log line provides no
benefit as it generates too much noise.

Bug: 131816428
Test: Builds, verified log message is not present

Change-Id: Ic657e12b050958cfb69467c83c4aab2bbd815d02