Fix fuzzer builds [#3].
Explicitly disabled fuzzer support on libsurfaceflinger. Generally (as
per aosp/954925) we simply disable LTO when SANITIZE_TARGET='fuzzer' is
specified, but in this case, libsurfaceflinger uses
-fwhole-program-vtables, which requires thinLTO. We explicitly disable
fuzzer support in this instance.

Bug: 121042685
Test: With all patches in the bug merged, build a fuzzer using
'SANITIZE_TARGET=fuzzer mmma <your_fuzzer>'.

This is the qt-dev-plus-aosp cherrypick of aosp/961628. See the AOSP
commit for further information.

Change-Id: Id6540dcc144113ede591e7326715ef8d1e01863d